node-notes-cli: update1
